requirements.txt
google-api-python-client==2.108.0
python-dotenv==1.0.0
pydantic==2.5.2
aiohttp==3.9.1
tenacity==8.2.3
youtube-transcript-api==0.6.2
openai==1.6.1
pytest==7.4.3
pytest-asyncio==0.23.2 